refactor: separate utils to files
